requirements.txt
asgiref==3.8.1
Django==5.2.1
django-constance==4.3.2
django-environ==0.12.0
sqlparse==0.5.3
